requirements.txt
anthropic~=0.25.8
PyYAML~=6.0.1
